Weekly_Tasks/Capstone Project/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
mysql-connector-python
pickle-mixin